refactor palette and auth
