+++ huevee-backend/routes/auth.js
@@ -1,38 +1,9 @@
-const express = require('express');
+import express from 'express';
+import { register, login } from '../controllers/authController.js';
+
 const router = express.Router();
-const pool = require('../db');
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
 
-router.post('/register', async (req, res) => {
-    const { username, email, password } = req.body;
-    const hash = await bcrypt.hash(password, 10);
-    try {
-        const result = await pool.query(
-            'INSERT INTO users (username, email, password_hash) VALUES ($1, $2, $3) RETURNING id',
-            [username, email, hash]
-        );
-        res.status(201).json({ userId: result.rows[0].id });
-    } catch (err) {
-        console.error('Register error:', err);
-        res.status(500).json({ error: 'Registration failed' });
-    }
-});
+router.post('/register', register);
+router.post('/login', login);
 
-router.post('/login', async (req, res) => {
-    const { email, password } = req.body;
-    try {
-        const result = await pool.query('SELECT * FROM users WHERE email = $1', [email]);
-        const user = result.rows[0];
-        if (!user || !(await bcrypt.compare(password, user.password_hash))) {
-            return res.status(401).json({ error: 'Invalid creds' });
-        }
-        const token = jwt.sign({ userId: user.id }, process.env.JWT_SECRET, { expiresIn: '1d' });
-        res.json({ token });
-    } catch (err) {
-        console.error('Login error:', err);
-        res.status(500).json({ error: 'Login failed' });
-    }
-});
-
-module.exports = router;
+export default router;
