+++ huevee-backend/server.js
@@ -1,13 +1,16 @@
-const express = require('express');
-const cors = require('cors');
-require('dotenv').config();
+import express from 'express';
+import paletteRoutes from './routes/palettes.js';
+import cors from 'cors';
+import dotenv from 'dotenv';
+import authRoutes from './routes/auth.js';
+
+dotenv.config();
 
 const app = express();
 app.use(cors());
 app.use(express.json());
 
 // Routes
-const paletteRoutes = require('./routes/palettes');
 app.use('/api/palettes', paletteRoutes);
 
 const PORT = process.env.PORT || 5000;
@@ -15,5 +18,4 @@ app.listen(PORT, () => {
     console.log(`Server is running on port ${PORT}`);
 });
 
-const authRoutes = require('./routes/auth');
 app.use('/api/auth', authRoutes);
